Match various functions

+++ src/4A3E0.h
@@ -6,6 +6,6 @@
 void func_800497E0(s16*, s32, u32, f32);
 void func_8004980C(u16, u8, u16);
 u8 func_80049890(u16);
-void func_800498A8(s32, s32, s32);
+void func_800498A8(u32, u8, u32);
 
 #endif // _4A3E0_H_
